[#53] Make WorkflowDispatch trigger extensible
By making it a class instead of an object, we force creating an object
which one day is expected to get some attributes like `inputs`. Thanks
to this, adding the attributes won't be a breaking change, and this
change is breaking. The sole feature of inputs in this case is
postponed and tracked in #64.