[#71] Rename CreateBranch wrapper to ActionCreateBranch
The `Action` prefix comes from the action's GitHub repo. If we wanted
to preserve just `CreateBranch`, it would mean being inconsistent with
the repo name. I figure it's better to follow the convention to make
the action better discoverable. Someone may not think that the name was
adjusted to not include `action`.