fix(workflows): run wrapper generation only on demand
When it was running for each branch besides master, it was making more
harm than good, e.g. it was impossible to create a nice git history in
PRs, or for Dependabot changes it was failing because Dependabot didn't
have proper permissions. Let's try running this workflow only on demand.