[#248] feat(wrappers): use v1 for bahmutov/npm-install
Using version up to patch number causes a need to update it frequently.
I discovered that this action does expose a major version but through
a branch, not a tag. Only tags are checked by the job that looks for
newer versions. This pontential room for improvement is tracked in #250.