[#129] feat(library): allow passing __FILE__ as source
There's a new feature in Kotlin scripting - `__FILE__` resolves to a
`File` object pointing to the script that it's being ran. It can be
leveraged to allow regenerating the script from any place the user
wishes, e.g. from IntelliJ or from `.github/workflows` directory. The
logic added here infers git root path and ensures the YAML is correctly
generated.

Co-authored-by: NikkyAI <git@nikky.moe>