[#407] feat(wrapper-generator)!: define typings in YAMLs
This change is needed to more conveniently manage the typings stored in
this repository, and to make it coherent with
https://github.com/krzema12/github-actions-typing/.

Some inconsistencies in namings of action generated types came up, so
this change is formally a breaking one for several actions. Previously
type names were defined in this repository for each input, and now the
wrapper generator infers it from the input name. In case it's desired
to have a custom name, a feature was added to be able to specify it in
the YAML.