[#208] feat(wrapper-generator): lock action versions (#546)
Instead of using the newest commit on some given action version
tag/branch, now the process of updating the commit and fetching
manifest is separated. Thanks to this, we get more stable `main` branch
as it won't stop building anymore due to some changes in the actions.
This library's maintainer has to call `./gradlew updateCommitHashes` to
change which commits are used to fetch action manifests.