chore(deps): update dependency gradle to v7.6 (#570)
It looks like the error was caused by a missing version
for the Kotlin Serialization plugin dependency in
buildSrc/builld.gradle.kts. The Kotlin BOM doesn't provide
the version for the Gradle plugins. In Gradle 7.6 the
embedded Kotlin version was bumped to 1.7.20 (from
1.6.20 I think), and I guess there's some incompatibility
between 1.7.20 (Gradle embedded Kotlin version) and
1.7.21 (the requested project version).

Or changing enforcedPlatform(kotlin("bom")) to
platform(kotlin("bom")) fixed it?