[#656] feat(library): support defining workflows not in Kotlin scripts (#657)
The goal of this change is to make inputs related to source Kotlin
scripts optional, so that e.g. workflow YAMLs can be written from
inside some build logic that doesn't sit in `.github/workflows`. In
this case, e.g. the consistency check doesn't make sense, or the
root git repo path has to be provided by hand. This makes the library
more independent on the usual layout, adding to its flexibility.