[#808] refactor(library)!: remove surplus overloads of 'uses' and 'run' (#810)
The only advantage of having two overloads of the methods is that
one can write run("command") - without specifying the name of the
step. It's very likely that we'll enforce using named arguments in the
whole API of this library (#807), so removing these overloads will only
help reducing maintenance effort of the library, and will still provide
equivalent functionality.