[#752] feat(library): add 'if' as alias for condition (#764)
In YAML, to set a condition for a step or a job, there's an 'if' property. Before this change,
this library had only 'condition' - not 'if' because it's a Kotlin keyword and to use it, one needs
backticks. After getting feedback from the community, we decided to add 'if' as an alias
for 'condition'.

Co-authored-by: Bens Keil <benkeil.me@googlemail.com>
Co-authored-by: Piotr Krzeminski <git@krzeminski.it>