[#941] refactor(automation): roughly split wrapper generation to separate module (#979)
This is the first from a series of changes aiming at being able to have
wrapper generation used on the client side. To be able to do it, we need
a module that is highly focused on wrapper generation, and doesn't
contain anything else. The separation is still not ideal, so other adjustment
changes are likely to follow.