test(library): fix integration tests using resources outside of module (#991)
It turns out that if the tests use resources outside of the 'resources' directory
designated for a given Gradle module, Gradle doesn't check if these external resources
changed to determine if the tests need to be rerun. In this case, it takes test results from
cache. This change makes Gradle check for changes in this extra place, invalidating cached
test results.