feat(abg): use typings for older version if available (#1204)
Fixes #1073.

If there's no typings for a given version in the catalog, typings for an
older version is tried to be fetched. Thanks to this, we cover a
scenario where usually the next version of some action doesn't need to
have typings defined immediately because the older version is fine. It
doesn't account for breaking changes where this approach will produce an
incorrect binding, however in practice it should be better than what we
have now.

It doesn't hold for action-hosted typings.