fix(automation): create new branch for each PR
It was a common problem that the automated PR was created with some
branch name, then the PR had to be closed, the branch wasn't removed,
and upon rerunning the workflow for PR creation, the PR reused the old
branch. The outcome was that instead of using the new contents in the
PR, the old contents was used because the branch was already existing -
GitHub reused the existing branch. By adding a timestamp to the branch
name, we effectively make the workflow create a new branch every time
the workflow is called.