test(abg): update generated file in place
Before this change, if the generated Kotlin file didn't match the
expected one, a new file was created under a different path. It was
then hard to just accept the changes i.e. use what was actually
produced as the expected file. This change makes the test edit the
file in place.