chore: remove release-docs workflow
This workflow is useless in such form because calling it on a -SNAPSHOT
version would publish docs with such version. In practice, I publish a
patch version to update just the docs, and 99% of the times, the
changes in the docs can wait for the next scheduled release. Removing
this workflow simplifies our GitHub scripts.