feat(dsl): allow providing callback to use workflow object (#1535)
Part of #1464.

It's the new API for using the `Workflow` domain object for whatever it
may be useful for, before it's serialized to YAML.
In version 3.0, the `workflow` DSL function will stop returning the
`Workflow` object, and this new API will be the only way to access the
object. Thanks to this, in most scripts, the workflow domain object
won't be printed to standard output anymore.