fix(docs): use lib's version for docs directly (#1551)
Closes #1550.

There's a problem with a breaking change in setuptools, which causes
issues when installing `mkdocs-markdownextradata-plugin`. This plugin
hasn't been touched for 3 years, so I doubt anyone will fix it. That's
why we're removing usage of it, it's pretty cheap.

We're going to stop using the lib's version in the docs as soon as we
stop bundling the action bindings, which is approx. in a month.