test(abg): specify typing in generation test (#1594)
Part of #1585.

It's needed because without this change, the generator infers there are
no typings for these test actions (the actions actually don't exist),
and in the new world it would only produce the `_Untyped` binding for
them. These tests actually check the path where the typings are
provided. For simplicity, we use `String` everywhere because there's
already a test that checks generation for other typings.