feat(abg)!: include untyped bindings (#1593)
Part of #1585.

For every binding JAR, include an untyped binding. If the typings are
available, additionally include a typed binding.

This change breaks clients that use actions without any typings. It
forces them to either add `_Untyped` to the class name, or add the
typings to the action or the typing catalog.