fix(abg): use the release date of the latest version for lastUpdated (#1612)
This makes the metadata file content more stable and allows to provide a checksum,
as Kotlin for example is quite unhappy if it finds no checksum or cannot successfully verify it.

Without this every request would produce a different file and thus also a different checksum file,
so the checksum would never match the file contents.