fix(server): configure endpoint for OpenTelemetry exporters (#1664)
Previously, the OpenTelemetry metric and log record exporters were
initialized without specifying an endpoint. This change sets the
endpoint configuration for both exporters to ensure they communicate
with the correct server.

Without this change, the default endpoint is `localhost`, which won't
resolve and thus part of the payload won't be published. This leads to a
 complete failure of metric/telemetry publishing, and it's the reason
 why many metrics are not making their way into Jaeger.

Making the change and running the docker repository locally proved to me
 that no more errors were showing up in the log. Next step is bringing
 it to production