fix(abg): handle non-numeric versions gracefully (#1668)
Fixes #1667.

In the logic, we were trying to parse version like `v2-beta` into an
integer. We could generally try to do it, but I think there are too
many possible variations to support them, so let's just support
versions that after removing the leading `v`, gives us an integer.