chore: rename githubToken to githubAuthToken (#1868)
Part of
https://github.com/typesafegithub/github-workflows-kt/issues/1854.

A "GitHub token" is, I think, a reserved name for a special kind of
token. With this change we'd like to express that the code can work with
various kinds of token that can be used to call GitHub's API, not
necessarily what is available under `GITHUB_TOKEN` and created by
GitHub, or not necessarily a Personal Access Token. In particulr, we're
working on switching to authenticating with a custom GitHub app where
we'll have Applicaiton Installation Token.