fix(library)!: allow running Kotlin-based step if multiple workflows in single file (#1870)
Resolves
https://github.com/typesafegithub/github-workflows-kt/issues/1811.

Includes target workflow name in an environment variable used to select
which Kotlin-based step to run. Before this fix, if there were multiple
workflows per a .main.kts file, there was a runtime error. Thanks to
this fix, things work fine as long as each workflow has a different
`targetFileName`.