fix(library): do not check if .git is a dir (#1900)
Closes
https://github.com/typesafegithub/github-workflows-kt/issues/1898.

`findGitDir` function is used to figure out where the script is run
from, so it's a critical function that is called upon each YAML
regeneration.

The check if it's a directory causes troubles when using git's
worktrees. `.git` is a symlink to the real `.git` dir, so it's not a
directory. Based on
https://github.com/typesafegithub/github-workflows-kt/issues/1898#issuecomment-2806850958.