fix(server): properly handle the lifecycle of ktor-client (#1913)
Fixes #1904

When using `ktor-client` you have to manage its lifecycle.

That means, if you create a `HttpClientEngine` manually, you have to
also `close()` it if it is no longer needed to free resources.
If you create the `HttpClient` with an `HttpClientEngineFactory`
instead, `ktor-client` is responsible to manage its lifecycle.

And if you create an `HttpClient`, you also have to `close()` it if it
is no longer needed to free resources, and also that the engine is
closed if it is managed by `ktor-client`.

Before this PR engines were created manually but never closed and
clients were created but never closed.
This lead to the problem that requests to the GitHub API became slower
and slower and eventually fail with GitHub closing or rejecting the
connections.
From evidence I guess that the GitHub servers have a limit of 200
concurrent connections for a given IP and when we reached that limit
over time because not closing the `ktor-client` objects, requests
failed.

With this PR we switch to give an engine factory to the client so that
`ktor-client` manages its lifecycle and properly close the http client
if it is no longer needed.
For the on-demand delayed date retrievals for versions, we now also
create a fresh client that is then closed after the request.