fix(server): treat failed binding generation as internal server error (#1924)
Before this PR, it was treated equally if a binding could not be found,
and if some exception happened during initial creating of the binding.
The exceptions were swallowed and the client received a 404 response.
If there was for example a compile error you already got internal server
error.
But if for example Kotlin Poet throws an illegal argument exception,
then this did not propagate properly

Now exceptions are logged and the client gets an internal server error
if some exception happened and only a 404 if indeed the action was not
found.

One such example is, if a typing has an `enum` with some option that
contains a colon.
This resulted in an identifier generated that contains a colon which
then produces an
exception due to the illegal identifier character.

---------

Co-authored-by: Piotr Krzemiński <3110813+krzema12@users.noreply.github.com>
Co-authored-by: Piotr Krzeminski <git@krzeminski.it>