refactor(server): do not leak cache in tests (#1933)
Thanks to this, the test aren't aware of an implementation detail of the
server: the cache.
We wrap the whole server config in an `appModule`, and inject a piece of
logic exposed by `maven-binding-builder` Gradle module. This piece is
the only thing we need to mock right now in the test. It makes the test
less brittle and couple with the implementation (although some necessary
coupling is still there).