test: use mock server for tests using GitHub API (#1972)
Part of
https://github.com/typesafegithub/github-workflows-kt/issues/1884.

Removes a dependency on ktor in the tests of GithubApi module. It's
needed because we're going to move away of using GitHub's API directly
via ktor, and use a library instead.
This change focuses only on refactoring the tests to use another mocking
mechanism - instead of using ktor's API, it exposes a real mock HTTP
server.