feat(library): add API for adding intended action version in comment (#2118)
Part of
https://github.com/typesafegithub/github-workflows-kt/issues/1691.

This change adds API to be able to add a comment with an intended action
version to the YAML. See an example:

```yaml
- id: 'step-0'
  name: 'Check out'
  uses: 'actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8' # v5.0.0
```

This follows the approach that Renovate uses, see
https://docs.renovatebot.com/modules/manager/github-actions/#digest-pinning-and-updating.
It's also a part of implementing [2025-04-27 Enhancements to referring
to actions by commit
hash](https://github.com/typesafegithub/design-docs/blob/main/2025-04-27%20Enhancements%20to%20referring%20to%20actions%20by%20commit%20hash.md).

This change needs to go as the very first one. It needs to land in a
library release, so that the server can generate action bindings using
the extended `RegularAction` base class.

ABI-wise, `RegularAction` is modified in a backward-compatible way
because classes inheriting from it are generated by the bindings server.
For `CustomAction`, all that is provided is source-level backward
compatibility which is IMO good enough.