chore(abg): lift creation of default HTTP client to generateBindings fun (#2164)
Part of
https://github.com/typesafegithub/github-workflows-kt/issues/2160.

I'd like to intercept calls to GitHub on the bindings server level, and
for this I'd like to use [HttpSend ktor
plugn](https://ktor.io/docs/client-http-send.html). To use it, I need to
inject the client augmented with this, and I want to do it from the
bindings server.