refactor(server): pass name and version from request to artifact generation (#2186)
This is a preliminary change making it easier to implement
https://github.com/typesafegithub/github-workflows-kt/issues/1691.

Even without the above project, it makes things more straightforward. In
particular, the artifact names happen to match what comes in the
request, but in general it doesn't have to be the case. We need to
encode certain extra information in Maven artifact's names and owners,
which leads to a mismatch where the GitHub action's name isn't equal to
the requested Maven artifact's name. This change recognizes it by using
different fields in these two circumstances.