feat(server): allow pinning action to commit hash (#2102)
Part of
https://github.com/typesafegithub/github-workflows-kt/issues/1691

DD:
https://github.com/typesafegithub/design-docs/blob/main/2025-04-27%20Enhancements%20to%20referring%20to%20actions%20by%20commit%20hash.md

This change implements the `commit_lenient` variant which doesn't
provide any validation on the commit hash (if the commit hash
corresponds to the tag referred to by the version). It just blindly uses
the given commit hash in `uses:` attribute, and the given full version
in the comment.